Services/Data/DashBoardDataService/src/main/java/com/sandata/lab/rest/dashboard/model/Staff.java
package com.sandata.lab.rest.dashboard.model;

import com.google.gson.annotations.SerializedName;

/**
 * Created by tom.dornseif on 4/11/2016.
 */
public class Staff {

    @SerializedName("FirstName")
    private String firstName;

    @SerializedName("LastName")
    private String lastName;

    @SerializedName("StaffId")
    private String staffId;

    @SerializedName("Value")
    private String value;

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getStaffId() {
        return staffId;
    }

    public void setStaffId(String staffId) {
        this.staffId = staffId;
    }

    public String getValue() {
        return value;
    }

    public void setValue(String value) {
        this.value = value;
    }
}
